+++ tests/testthat/test_cooccurrences.R
@@ -2,7 +2,7 @@ library(testthat)
 library(data.table)
 library(topicanalysis)
 
-testthat::context("zoo")
+testthat::context("cooccurrences")
 
 test_that(
   "cooccurrences plausibility checks",
@@ -51,4 +51,43 @@ test_that(
 
   }
 )
-      
+
+
+test_that(
+  "check cooccurrences results against docs()-method (without renumbering)",
+  {
+    data(BE_lda)
+    BE <- Topicanalysis$new(BE_lda)
+    BE_coocs <- cooccurrences(BE_lda, k = 5L)
+    for (i in 1:5){
+      expect_equal(
+        BE_coocs[i][["count_coi"]],
+        length(BE$docs(x = BE_coocs[i][["a"]], y = BE_coocs[i][["b"]], n = 5L))
+      )
+    }
+  }
+)
+
+
+test_that(
+  "check cooccurrences results against docs()-method (with renumbering, without subsetting)",
+  {
+    data(BE_lda)
+    top <- topicmodels::topics(BE_lda, k = 5L)
+    benchmark <- table(apply(top, 2, function(x) all(any(1:5 %in% x), any(6:10 %in% x))))[["TRUE"]]
+    
+    BE <- cooccurrences(BE_lda, k = 5L)
+    BE <- Topicanalysis$new(BE_lda)
+    
+    n_docs_method <- length(BE$docs(x = 1:5, y = 6:10, n = 5L))
+    
+    expect_identical(benchmark, n_docs_method)
+    
+    renumber_vec <- 1L:BE_lda@k
+    renumber_vec[1:5] <- 1L
+    renumber_vec[6:10] <- 6L
+    coocs <- cooccurrences(BE_lda, k = 5L, renumber = renumber_vec)
+    
+    expect_equal(coocs[a == 1][b == 6][["count_coi"]], n_docs_method)
+  }
+)